removed most dead code
