--- src/common/workspaces.rs
@@ -1,155 +1,0 @@
-use i3_ipc::{reply::Workspace, Connect, I3};
-
-use crate::CONFIG;
-
-use super::constraint::{Constraint, Constraints};
-use super::{groups, moves, name, outputs};
-
-fn get_ws() -> Vec<Workspace> {
-	let mut i3 = I3::connect().unwrap();
-	i3.get_workspaces().unwrap()
-}
-
-pub fn maintenance() {
-	let mut workspaces = get_ws();
-
-	workspaces.sort_by(|w1, w2| w1.num.cmp(&w2.num));
-
-	let mut i: i32 = 0;
-	for ws in workspaces {
-		i += 1;
-		moves::reorder(ws, i);
-	}
-}
-
-pub fn get(constraints: Constraints, reverse: bool) -> Vec<Workspace> {
-	let mut workspaces = get_ws();
-
-	if reverse {
-		workspaces.sort_by(|w1, w2| w2.num.cmp(&w1.num));
-	} else {
-		workspaces.sort_by(|w1, w2| w1.num.cmp(&w2.num));
-	}
-
-	if constraints.contains(Constraint::None) {
-		return workspaces;
-	}
-
-	workspaces
-		.iter()
-		.filter(|ws| {
-			if constraints.contains(Constraint::Output) && constraints.output != "none" {
-				if constraints.output != ws.output {
-					return false;
-				}
-			}
-
-			if constraints.contains(Constraint::AllowUrgent) && ws.urgent {
-				return true;
-			}
-
-			if constraints.contains(Constraint::Focused) {
-				if !ws.focused {
-					return false;
-				}
-			}
-
-			if constraints.contains(Constraint::Visible) {
-				if !ws.visible {
-					return false;
-				}
-			} else if constraints.contains(Constraint::Hidden) {
-				if ws.visible {
-					return false;
-				}
-				if ws.focused {
-					return false;
-				}
-			}
-
-			let ws_group = name::group(ws.name.as_ref());
-			if ws_group.len() == 0 && constraints.contains(Constraint::NoGroup) {
-				let output = if constraints.output != "none" {
-					constraints.output.clone()
-				} else {
-					outputs::focused()
-				};
-				let active_groups = groups::active(output.to_owned());
-				let available_groups = groups::available_output(output);
-				return !CONFIG.focus.hide_unassigned_workspaces
-					|| active_groups == available_groups;
-			}
-
-			if constraints.contains(Constraint::Group) {
-				let output = if constraints.output != "none" {
-					constraints.output.clone()
-				} else {
-					outputs::focused()
-				};
-				let active_groups = groups::active(output);
-				if active_groups.len() > 0 {
-					return active_groups.contains(&ws_group);
-				}
-			}
-
-			return true;
-		})
-		.cloned()
-		.collect::<Vec<Workspace>>()
-}
-
-pub fn first(constraints: Constraints) -> Workspace {
-	get(constraints, false).first().unwrap().to_owned()
-}
-
-pub fn last(constraints: Constraints) -> Workspace {
-	get(constraints, true).first().unwrap().to_owned()
-}
-
-pub fn visible_or_focused(output: &str) -> Workspace {
-	let workspaces = get_ws();
-
-	workspaces
-		.iter()
-		.find(|ws| ws.output == output && (ws.focused || ws.visible))
-		.unwrap()
-		.to_owned()
-}
-
-pub fn visible() -> Vec<Workspace> {
-	let workspaces = get_ws();
-
-	workspaces
-		.iter()
-		.filter(|ws| ws.visible || ws.focused)
-		.cloned()
-		.collect::<Vec<Workspace>>()
-}
-
-pub fn focused() -> Workspace {
-	let workspaces = get_ws();
-
-	workspaces.iter().find(|ws| ws.focused).unwrap().to_owned()
-}
-
-pub fn by_name(name: String) -> Option<Workspace> {
-	let workspaces = get_ws();
-
-	let ws = workspaces.iter().find(|ws| ws.name == name);
-
-	match ws {
-		Some(w) => Some(w.to_owned()),
-		None => None,
-	}
-}
-
-pub fn by_num(n: i32) -> Option<Workspace> {
-	let workspaces = get_ws();
-
-	let ws = workspaces.iter().find(|ws| ws.num == n);
-
-	match ws {
-		Some(w) => Some(w.to_owned()),
-		None => None,
-	}
-}
