--- src/polybar/presets/mod.rs
@@ -1,5 +1,0 @@
-mod show_hidden_toggle;
-
-use super::{Actions, Format, Label};
-
-pub use show_hidden_toggle::{generate as show_hidden_toggle, ToggleMode};
